+++ server/api/blog/controller/ex_photo_album.ctl.go
@@ -41,7 +41,7 @@ func (s *PhotoAlbumController) FindPhotoAlbumDetailsList(c *gin.Context) {
 		List:     list,
 		Total:    total,
 		Page:     0,
-		PageSize: len(list),
+		PageSize: int64(len(list)),
 	})
 }
